workaround: map TUS ID to media ID
tus does not yet have proper protocol support for returning metadata
after successful upload. So while waiting for that to happen we kludge
around it by creating a mapping table which a client may use to obtain
the final media ID after a successful tusd upload.

It's not pretty, but it'll be easy to remove.